+++ src/bk-user/bkuser/apis/open_v3/views/__init__.py
@@ -15,6 +15,7 @@
 # We undertake not to change the open source license (MIT license) applicable
 # to the current version of the project delivered to anyone in the future.
 
-from .tenant import TenantListApi, TenantUserDisplayNameListApi, TenantUserInfoRetrieveApi
+from .tenant import TenantListApi
+from .user import TenantUserDisplayNameListApi, TenantUserRetrieveApi
 
-__all__ = ["TenantListApi", "TenantUserDisplayNameListApi", "TenantUserInfoRetrieveApi"]
+__all__ = ["TenantListApi", "TenantUserDisplayNameListApi", "TenantUserRetrieveApi"]
