+++ src/bk-user/bkuser/apis/open_v3/views/tenant.py
@@ -20,13 +20,8 @@ from drf_yasg.utils import swagger_auto_schema
 from rest_framework import generics, status
 
 from bkuser.apis.open_v3.mixins import OpenApiCommonMixin
-from bkuser.apis.open_v3.serializers.tenant import (
-    TenantListOutputSLZ,
-    TenantUserDisplayNameListInputSLZ,
-    TenantUserDisplayNameListOutputSLZ,
-    TenantUserInfoRetrieveOutputSLZ,
-)
-from bkuser.apps.tenant.models import Tenant, TenantUser
+from bkuser.apis.open_v3.serializers.tenant import TenantListOutputSLZ
+from bkuser.apps.tenant.models import Tenant
 
 logger = logging.getLogger(__name__)
 
@@ -44,58 +39,3 @@ class TenantListApi(OpenApiCommonMixin, generics.ListAPIView):
     )
     def get(self, request, *args, **kwargs):
         return self.list(request, *args, **kwargs)
-
-
-class TenantUserDisplayNameListApi(OpenApiCommonMixin, generics.ListAPIView):
-    """
-    批量根据用户 bk_username 获取用户展示名
-    TODO: 性能较高，只查询所需字段，后续开发 DisplayName 支持表达式配置时添加 Cache 方案
-    """
-
-    pagination_class = None
-
-    serializer_class = TenantUserDisplayNameListOutputSLZ
-
-    def get_queryset(self):
-        slz = TenantUserDisplayNameListInputSLZ(data=self.request.query_params)
-        slz.is_valid(raise_exception=True)
-        data = slz.validated_data
-
-        # TODO: 由于目前 DisplayName 渲染只与 full_name 相关，所以只查询 full_name
-        # 后续支持表达式，则需要查询表达式可配置的所有字段
-        return (
-            TenantUser.objects.filter(id__in=data["bk_usernames"])
-            .select_related("data_source_user")
-            .only("id", "data_source_user__full_name")
-        )
-
-    @swagger_auto_schema(
-        tags=["open_v3.tenant"],
-        operation_id="batch_query_user_display_name",
-        operation_description="批量查询用户展示名",
-        query_serializer=TenantUserDisplayNameListInputSLZ(),
-        responses={status.HTTP_200_OK: TenantUserDisplayNameListOutputSLZ(many=True)},
-    )
-    def get(self, request, *args, **kwargs):
-        return self.list(request, *args, **kwargs)
-
-
-class TenantUserInfoRetrieveApi(OpenApiCommonMixin, generics.RetrieveAPIView):
-    lookup_url_kwarg = "tenant_user_id"
-    serializer_class = TenantUserInfoRetrieveOutputSLZ
-
-    def get_queryset(self):
-        # TODO: 由于目前 DisplayName 渲染只与 full_name 相关，所以只查询 full_name
-        # 后续支持表达式，则需要查询表达式可配置的所有字段
-        return TenantUser.objects.select_related("data_source_user").only(
-            "id", "data_source_user__full_name", "time_zone", "language", "tenant_id"
-        )
-
-    @swagger_auto_schema(
-        tags=["open_v3.tenant"],
-        operation_id="query_user_info",
-        operation_description="查询用户详情",
-        responses={status.HTTP_200_OK: TenantUserInfoRetrieveOutputSLZ()},
-    )
-    def get(self, request, *args, **kwargs):
-        return self.retrieve(request, *args, **kwargs)
