+++ src/bk-user/support-files/apidocs/zh/batch_query_user_display_name.md
@@ -4,9 +4,9 @@
 
 ### 输入参数
 
-| 参数名称         | 参数类型   | 必选 | 描述                       |
-|--------------|--------|----|--------------------------|
-| bk_usernames | string | 是  | 蓝鲸唯一标识，多个以逗号分隔，限制数量为 100 |
+| 参数名称         | 参数类型   | 必选 | 描述                         |
+|--------------|--------|----|----------------------------|
+| bk_usernames | string | 是  | 蓝鲸用户唯一标识，多个以逗号分隔，限制数量为 100 |
 
 ### 请求示例
 
@@ -34,10 +34,10 @@ bk_usernames=7idwx3b7nzk6xigs,0wngfim3uzhadh1w
 
 ### 响应参数说明
 
-| 参数名称         | 参数类型   | 描述     |
-|--------------|--------|--------|
-| bk_username  | string | 蓝鲸唯一标识 |
-| display_name | string | 用户展示名  |
+| 参数名称         | 参数类型   | 描述       |
+|--------------|--------|----------|
+| bk_username  | string | 蓝鲸用户唯一标识 |
+| display_name | string | 用户展示名    |
 
 ### 状态码非 200 的响应示例
