+++ src/bk-user/support-files/resources.yaml
@@ -59,8 +59,8 @@ paths:
 
   /api/v3/open/tenant/users/{bk_username}/:
     get:
-      operationId: query_user_info
-      description: 查询用户详情
+      operationId: retrieve_user
+      description: 查询用户信息
       tags: []
       responses:
         default:
